+++ plugins/nf-schema/src/main/nextflow/validation/help/HelpObserver.groovy
@@ -1,4 +1,4 @@
-package nextflow.validation
+package nextflow.validation.help
 
 import groovy.util.logging.Slf4j
 
@@ -11,7 +11,7 @@ import nextflow.validation.help.HelpMessage
 import nextflow.validation.config.ValidationConfig
 
 @Slf4j
-class ValidationObserver implements TraceObserver {
+class HelpObserver implements TraceObserver {
     
     @Override
     void onFlowCreate(Session session) {
