+++ plugins/nf-schema/src/resources/META-INF/extensions.idx
@@ -1,2 +1,2 @@
 nextflow.validation.ValidationExtension
-nextflow.validation.ValidationObserverFactory
+nextflow.validation.help.HelpObserverFactory
